+++ builtin_shaders-2017.4.15f1/CGIncludes/UnityPBSLighting.cginc
@@ -65,7 +65,7 @@ inline half3 BRDF_Unity_Indirect (half3 baseColor, half3 specColor, half oneMinu
 struct SurfaceOutputStandard
 {
     fixed3 Albedo;      // base (diffuse or specular) color
-    fixed3 Normal;      // tangent space normal, if written
+    float3 Normal;      // tangent space normal, if written
     half3 Emission;
     half Metallic;      // 0=non-metal, 1=metal
     // Smoothness is the user facing name, it should be perceptual smoothness but user should not have to deal with it.
@@ -75,7 +75,7 @@ struct SurfaceOutputStandard
     fixed Alpha;        // alpha for transparencies
 };
 
-inline half4 LightingStandard (SurfaceOutputStandard s, half3 viewDir, UnityGI gi)
+inline half4 LightingStandard (SurfaceOutputStandard s, float3 viewDir, UnityGI gi)
 {
     s.Normal = normalize(s.Normal);
 
@@ -93,7 +93,7 @@ inline half4 LightingStandard (SurfaceOutputStandard s, half3 viewDir, UnityGI g
     return c;
 }
 
-inline half4 LightingStandard_Deferred (SurfaceOutputStandard s, half3 viewDir, UnityGI gi, out half4 outGBuffer0, out half4 outGBuffer1, out half4 outGBuffer2)
+inline half4 LightingStandard_Deferred (SurfaceOutputStandard s, float3 viewDir, UnityGI gi, out half4 outGBuffer0, out half4 outGBuffer1, out half4 outGBuffer2)
 {
     half oneMinusReflectivity;
     half3 specColor;
@@ -134,14 +134,14 @@ struct SurfaceOutputStandardSpecular
 {
     fixed3 Albedo;      // diffuse color
     fixed3 Specular;    // specular color
-    fixed3 Normal;      // tangent space normal, if written
+    float3 Normal;      // tangent space normal, if written
     half3 Emission;
     half Smoothness;    // 0=rough, 1=smooth
     half Occlusion;     // occlusion (default 1)
     fixed Alpha;        // alpha for transparencies
 };
 
-inline half4 LightingStandardSpecular (SurfaceOutputStandardSpecular s, half3 viewDir, UnityGI gi)
+inline half4 LightingStandardSpecular (SurfaceOutputStandardSpecular s, float3 viewDir, UnityGI gi)
 {
     s.Normal = normalize(s.Normal);
 
@@ -159,7 +159,7 @@ inline half4 LightingStandardSpecular (SurfaceOutputStandardSpecular s, half3 vi
     return c;
 }
 
-inline half4 LightingStandardSpecular_Deferred (SurfaceOutputStandardSpecular s, half3 viewDir, UnityGI gi, out half4 outGBuffer0, out half4 outGBuffer1, out half4 outGBuffer2)
+inline half4 LightingStandardSpecular_Deferred (SurfaceOutputStandardSpecular s, float3 viewDir, UnityGI gi, out half4 outGBuffer0, out half4 outGBuffer1, out half4 outGBuffer2)
 {
     // energy conservation
     half oneMinusReflectivity;
